3: Move parser code into its own package
